+++ migrations/0012_class_c_node.sql
@@ -1,0 +1,7 @@
+-- +migrate Up
+alter table node
+	add column is_class_c boolean not null default false;
+
+-- +migrate Down
+alter table node
+	drop column is_class_c;