+++ docs/changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.4.0
+
+**Features & changes:**
+
+* Class-C support. When adding / changing nodes, tick the Class-C checkbox
+  to enable Class-C support for a given node.
+* Application ID added to application overview (as it is needed to subscribe
+  to MQTT topics).
+* Nodes are now sorted by name.
+
+**Note:** For Class-C functionality, upgrade LoRa Server to 0.14.0 or above.
+
 ## 0.3.0
 
 **This release contains changes that are not backwards compatible!**